Refactor consent.tsx to render App
